+++ src/LSS.h
@@ -11,45 +11,45 @@ class LSS
 {
 	int m, n, Q, mstep;
 	SolverFactory *solver;
-	double **** trajectory;
-	double ****b;
-	double ***eq;
-	double **rho, **u, **v;
-	double *cx, *cy, *w;
-	double p, omega, u0;
+	float **** trajectory;
+	float ****b;
+	float ***eq;
+	float **rho, **u, **v;
+	float *cx, *cy, *w;
+	float p, omega, u0;
 	
 public:
 	LSS();
 	~LSS();
 
 	//Friend functions
-	friend void Obj(double p, double u0, int m, int n, double *cx, double *cy, double *w, double **rho, double **u, double **v, double omega, double ***feq, double*** fin, double*** fout, double *J);
+	friend void Obj(float p, float u0, int m, int n, float *cx, float *cy, float *w, float **rho, float **u, float **v, float omega, float ***feq, float*** fin, float*** fout, float *J);
 
 	//Baseline solvers
-	void CreateCase(double u0_, int tsteps_, int m_, int mx) { solver = SolverFactory::Create(SolverFactory::LB,u0_,tsteps_,m_,mx); } //One solver available
-	void SetCase(double s, int t, int m, int mx);
+	void CreateCase(float u0_, int tsteps_, int m_, int mx) { solver = SolverFactory::Create(SolverFactory::LB,u0_,tsteps_,m_,mx); } //One solver available
+	void SetCase(float s, int t, int m, int mx);
 	void SolveCase() { solver->Exectue(); }
-	double GetObjectiveFuntion() { return solver->GetObjectiveFunction(); }
-	double**** GetTrajectory() { return solver->GetTrajectory(); }
+	float GetObjectiveFuntion() { return solver->GetObjectiveFunction(); }
+	float**** GetTrajectory() { return solver->GetTrajectory(); }
 
 	//Auxiliary functions
-	double Norm(double *, int);
-	double VectorNorm(double ****);
-	double ScalProduct(int, double*, double*);
-	double ScalarProduct(double****, double****);
+	float Norm(float *, int);
+	float VectorNorm(float ****);
+	float ScalProduct(int, float*, float*);
+	float ScalarProduct(float****, float****);
 
 	//LSS functions
-	void AssemblyArray(double****,double****);
-	double CalculateSensitivity(double****);
+	void AssemblyArray(float****,float****);
+	float CalculateSensitivity(float****);
 	void CreateRHSVector();
 	void GetMacroscopic(int);
-	void Precond(int, double *, double*, double*);
-	void Preconditioner(double****, double****);
+	void Precond(int, float *, float*, float*);
+	void Preconditioner(float****, float****);
 	void Preprocess();
-	void Solve(void(*mult)(double*, double*, double*, int),double*,double*);
+	void Solve(void(*mult)(float*, float*, float*, int),float*,float*);
 	void SolveKKT();
 
 	//Export functions
-	void Export(double****);
+	void Export(float****);
 };
